[Patch] Move error func to own file, add conversion to channel_outbound
